file: Support punching holes for write zero
On Linux, use fallocate() to punch holes as a more efficient way
of writing zeroes.  If hole punching is not allowed, or if we
can't use fallocate (whether because this is not Linux, or because
the file system on Linux doesn't support it), gracefully fall
back to the write method.  If wdelayms is set and a fallback occurs,
we end up sleeping at least twice as long; oh well.

Signed-off-by: Eric Blake <eblake@redhat.com>