file: Optimize writing zeroes without holes
On newer Linux, fallocate() can more efficiently write zeroes
without punching holes.  When this mode is supported, use it
instead of falling back to write (while still gracefully falling
back to the write method if it fails with EOPNOTSUPP).

Also fix a typo in an error message in the previous commit.

Signed-off-by: Eric Blake <eblake@redhat.com>