plugins: Add new nbdkit_set_error() utility function
The last patch fixed the plugin interface to not use a bogus value
of errno from other language bindings, but hard-coding things to
EIO is not nice either.

The solution is to expose an explicit utility function for setting
the preferred error value, while still falling back to errno for
backwards compatibility.  The saved error code has to be
thread-safe for use in multithreaded plugins.

Note that setting an error value but then returning success from
the plugin has no impact; the error value is only consulted when
the plugin reports failure, and is intentionally wiped before
each callback.

Signed-off-by: Eric Blake <eblake@redhat.com>