build: Kill build on POD error
I recently introduced (and fixed) a POD error that escaped detection
because make did not die.  As penance, I'm improving the build
machinery, borrowing ideas from libvirt.

The pod2man shipped in perl-podlators 4.09 (Fedora 25) defaults to
'--errors=die' which prints the errors and exits with a non-zero
status on any problem; but for back-compat, if --stderr is used,
this overrides the default, and ends up still printing warnings
to stderr but with no change to exit status (effectively ignoring
errors).  Omitting --stderr is thus saner behavior; grepping for
'POD ERRORS' will never succeed, but doesn't matter because the
pod2man call itself fails.

The pod2man shipped with perl 5.10 (RHEL 6 timeframe) lacks
'--errors=die' but has '--stderr'; there the use of '--stderr'
prints errors to stderr instead of to a 'POD ERRORS' section of
the output, but does not affect exit status.  Omitting --stderr
means that you no longer see the error, but adding the grep for
'POD ERRORS' will at least halt the build.

The pod2man shipped with perl 5.8 (RHEL 5 timeframe) lacks '--stderr',
and always spits errors to stderr with a 0 exit status; it never
produces a 'POD ERRORS' section in the output.  Lame, but it merely
means that someone developing on RHEL 5 won't catch pod errors until
they rebuild on a newer platform.

Signed-off-by: Eric Blake <eblake@redhat.com>