vddk: Extend the plugin to support access to remote servers.
This adds support for accessing remote VMware ESXi or vCenter servers
using VDDK.

Some other small changes are made:

 - Display transport mode in debug output.

 - Use ‘static’ scope for plugin variables.

 - Update the documentation.

Note there is a user-visible change: The file=... parameter must now
be specified as an absolute path for local files.  This is because
when using a remote server, this path is a "[datastore] path" on the
remote server which must not be altered by nbdkit.