Introduce top-level ./nbdkit for running from the source directory.
This rather large change starts by introducing a script in the top
build directory for running nbdkit from the source directory instead
of having to install it.

I also reworked the tests to use this script.  Instead of calling
../src/nbdkit, then just use ‘nbdkit’ and $PATH is set up to run the
script in the top build directory.  Also there is no longer any need
to munge the plugin name in tests since the script does that.

The above changes break valgrinding, but in fact that leads to a
simplification: The top level script takes an environment variable
(‘NBDKIT_VALGRIND=1’) to enable valgrinding, and there is no need to
use the ‘$VG’ hack in the tests.  In addition this means we can remove
some wrapper scripts around C tests which only had the purpose of
expanding the ‘$VG’ variable.