Wait up to 5 seconds for running threads to complete before exiting.
There is a race condition where nbdkit exits (calling plugin_cleanup()
which unloads the plugin), while the plugin is in its plugin.close()
function in another thread.

This keeps track of the number of running threads and waits up to 5
seconds for the running threads to exit before exiting the main
program.

For further analysis, see also:
https://www.redhat.com/archives/libguestfs/2017-September/msg00226.html