maint: Add NBDKIT_GDB support to in-tree nbdkit script
During development, it is handy to use the ./nbdkit script to
automatically pick up the just-built binaries without having
to install things; but using gdb on a script does not work.
Similar to how we already support $NBDKIT_VALGRIND, add support
for $NBDKIT_GDB; and also make the script robust to a value
of $prefix inherited from the environment.

Signed-off-by: Eric Blake <eblake@redhat.com>