Avoid race conditions when nbdkit exits.
There are several race conditions where nbdkit exits (calling
plugin_cleanup() which unloads the plugin), while the plugin is either
in a callback or in plugin.close().

Avoid these by:

(1) Taking a shared lock around all plugin callbacks.

(2) Taking the same as an exclusive lock in plugin_cleanup.

This delays plugin_cleanup until all callbacks have finished.

(3) Removing a few unnecessary ‘assert (dl)’.

This is necessary because the plugin can now be unloaded while holding
locks.  ‘dl’ is not needed to lock or unlock the plugin so asserting
it is useless.

(4) Don't call plugin.close on the quit path.

Another thread could be unloading the plugin so we cannot call
the .close callback.