connections: Report mid-message EOF as fatal
If we encounter any read error in the middle of the message
(including the NBD_CMD_WRITE payload), we are no longer in sync
with the client and should not try to read anything further.
For mid-message EOF, the problem wasn't too bad (the next
iteration of the loop would also see EOF); but if it is some
other error (perhaps a transient EIO due to reading from a
flaky block device), not marking the connection failed could
result in us trying to treat the tail of the previous partial
payload as further commands.  Usually a magic number mismatch
would flag the problem, but we should never base our behavior
on the contents of that random payload.

Signed-off-by: Eric Blake <eblake@redhat.com>