connections: Simplify handle_request()
_handle_request() was never returning anything but 0; and had to
communicate the error value through a parameter by reference.
As there is no return value that a plugin can use to declare the
connection useless, we can consolidate this into just returning
the error value.  Meanwhile, we don't need a wrapper layer of
handle_request() just to grab locks; so shuffle the naming a
bit to reduce the lines of code.

Signed-off-by: Eric Blake <eblake@redhat.com>