sockets: Fix lifetime of thread_data
It is never a wise idea to pass stack-allocated storage to
another thread during pthread_create() unless you can guarantee
that the new thread will complete prior to the calling thread
returning and ending the lifetime of that storage.  We were
violating this, which I managed to detect as a SEGV while
running test-socket-activation in a loop: a detached child
thread was using threadlocal_set_sockaddr() with parameters
pointing into obsolete thread_data, at the same time the main
thread was trying to call exit().

Signed-off-by: Eric Blake <eblake@redhat.com>