tests: Test parallel nbd behavior
Copy the paradigm of test-parallel-file.sh to further test that the
nbd forwarder can handle out-of-order requests.  Once again, the
delay values I chose requires at least 5 seconds to complete; I don't
know if it is worth trying to fine-tune it to run faster while still
being robust in the presence of a heavy load.

With multiple nbdkit processes active during the test, it gets hairy
to write a nested --run that properly quotes the arguments needed for
two automatic Unix socket names.  So this test instead resorts to
running the file plugin using --exit-with-parent within a subshell.

Signed-off-by: Eric Blake <eblake@redhat.com>