tests: Test parallel file behavior
Codify the testing procedure mentioned during the addition of --threads
into a full-blown member of 'make check'.  With the delay values I
chose, the test requires at least 5 seconds to complete; I don't know
if it is worth trying to fine-tune it to run faster while still being
robust in the presence of a heavy load.

This requires a slight refactoring of the Makefile.am to provide
file-data to more consumers.

Signed-off-by: Eric Blake <eblake@redhat.com>