tests: Prepare for running multiple nbdkits in one test
The obvious way to test the nbd plugin is to wrap yet another
nbdkit server.  The testsuite already has nice cleanup of an
nbdkit server, which we want to reuse while multiple servers
are running.  So rework global variables to instead belong to
a struct which is allocated per invocation, with the cleanup
function destroying servers in LIFO order.

To minimize changes to existing tests, existing global variables
are changed to track the details of the most-recently created
server.  A new global sock variable will also prove useful.

Signed-off-by: Eric Blake <eblake@redhat.com>