tests: Test nbd forwarder plugin
I'd be remiss if I didn't enhance the testsuite to exercise the
nbd plugin (after all, the best way to make sure something doesn't
regress is to cover it in the testsuite).  Borrows heavily from
test-file, with the main difference being that I wrap three nbdkit
processes instead of one.

Signed-off-by: Eric Blake <eblake@redhat.com>