nbd: Enable parallel handling
The final step in making the nbd forwarder fully parallel is to
track a list of pending transactions, rather than a single
transaction.  The list is only ever manipulated with the lock
taken, so all points that use the list see a consistent state;

If we get a response that does not have a cookie from our list,
we have no choice but to consider the server broken and declare
the connection dead.  If we get an ESHUTDOWN from the server,
we rely on nbdkit's main handling to eventually shut us down,
as we do not want to interrupt any other pending in-flight
operations.  Once our reader loop quits, we make sure to mop up
any remaining in-flight operations.

Signed-off-by: Eric Blake <eblake@redhat.com>