nbd: Properly mop up stranded requests
Nothing in the second loop of nbd_reader() was ever setting
quit; the only way the loop could end was via SEGV for
dereferencing NULL or abort() when attempting to write to
an invalid fd.  Finish the implementation to properly mop
up stranded in-flight requests.

The testsuite did not exercise the case of quitting
nbdkit while a transaction was in flight, but was relying
on SIGTERM exiting the entire process before all threads
were reaped; otherwise, I might have spotted this sooner.

Signed-off-by: Eric Blake <eblake@redhat.com>