file: Diagnose a missing file earlier
If a user typos the filename for the file plugin, they do not
find out about it until later when the first client gets an
abrupt hangup, when file_open() fails to open things.  Better
is to avoid starting a server at all if the file doesn't exist
yet.

While at it, avoid a memory leak if the caller passes a file=
argument more than once.

Signed-off-by: Eric Blake <eblake@redhat.com>