tests: Make parallel tests more robust
nbdkit simulates the NBD FUA flag by invoking flush() after
write.  Under a heavily-loaded system, especially when both
parallel tests are hammering the same file, this flush can
be expensive enough that the write lasts longer than the
corresponding read, in spite of the file plugin having a
shorter wdelay.  (Note that the tests write idempotent data
into the file, so there is no risk of the file contents
changing).  Switch the tests to favor reads as the second
but short operation, so that the test is not as sensitive
to the timing taken for a flush after write.

Signed-off-by: Eric Blake <eblake@redhat.com>