Add a pkg-config/pkgconf file.
Unlike a regular library there is no need to use the pkg-config file
to set $CFLAGS or $LIBS.  However a pkg-config file is still useful
for a few reasons:

(1) Can use PKG_CHECK_MODULES([NBDKIT], [nbdkit >= 1.1.25]) to check
for existence and minimum version of the nbdkit plugin development
environment at configure time, which is useful for out of tree
plugins.

(2) Can query the plugindir at compile time:

  pkgconf nbdkit --variable=plugindir

(3) Could add $CFLAGS or $LIBS in future (although that might be
considered an API break so there would need to be a very good reason
for it).