protocol: Split flags from cmd field in requests
Since NBD is a big-endian protocol, the upstream spec was able to
repurpose a 32-bit field with flags starting at (1<<16) OR'd into
the command into two 16-bit fields (flags first, starting at 1<<0,
then the command) for ease of documentation.  Matching that split
in our code base will also make it easier to implement smarter FUA
flag support.  This addresses one of the TODO in the nbd plugin.

Signed-off-by: Eric Blake <eblake@redhat.com>