Refactor plugin_* functions into a backend struct.
Introduce the concept of a backend.  Currently the only type of
backend is a plugin, and there can only be one of them.  Instead of
calling functions like ‘plugin_pwrite’ you call the backend method
‘backend->pwrite (backend, ...)’.

The change is largely mechanical.  I was able to remove ‘assert (dl)’
statements throughout since we can now prove they will never be
called.

Note this does not lift the restriction of one plugin per server, and
it can *never* do that because plugins can use global variables.
Message-Id: <20180117205356.8699-3-rjones@redhat.com>