backend: Pass flags argument through
Although our plugin interface has to support older users that
didn't take a flag parameter, we want to allow newer plugins
that can honor flags, as well as insist that all filters take a
flag argument.  This will allow future expansion to conditionally
honor any flags passed over the NBD protocol; with immediate
plans to support NBD_CMD_FLAG_FUA.  So, make the backend
interface always pass flags for actions corresponding to NBD
wire commands.  The .zero interface is currently the only user
of a flag; note that our NBDKIT_FLAG_MAY_TRIM preserves our old
internal sense, which is intentionally opposite of the NBD
protocol NBD_CMD_FLAG_NO_HOLE, and that our flag values are not
the same as the on-the-wire flag values.

Signed-off-by: Eric Blake <eblake@redhat.com>