plugins: Move FUA fallback to plugins
As long as the plugins couldn't directly handle Forced Unit Access
(FUA), it made sense to have a common fallback in handle_request().
But upcoming patches will allow plugins to handle FUA, at which
point the fallback is easier to implement on a per-command basis,
via the just-added flags parameter in backend functions.
Furthermore, the upcoming addition of filters means that we do not
want to call .flush at the wrong layer of the backend stack, so it
makes sense to have the fallback as low as possible.

The NBD spec says that we must tolerate a client sending
NBD_CMD_FLAG_FUA on any command (due to historical behavior
of at least qemu sending it on READ), but that it only has
to have defined semantics on commands that can cause write
actions.  So, we only pass the FUA flag through on WRITE,
WRITE_ZEROES, and TRIM.

Note that validate_request() already ensured that that we are
not calling a write command if conn->readonly; so we no longer
have to check that.  For now, FUA support is synonymous with
.can_flush support, because we are still implementing FUA via
the fallback to .flush; but future patches will split the
two conditions as part of wiring up further FUA support.  It
also means that we can assert that if we did not advertise
FUA support to the client, then the client should not be
requesting FUA; and therefore if the plugins layer sees the
FUA flag, we must have a .flush callback.

Signed-off-by: Eric Blake <eblake@redhat.com>