include: Properly decorate attributes in public headers
gcc and clang intentionally support __decorated__ forms of all
attribute arguments (in the namespace reserved to the compiler
implementation), precisely so that any public header does
not risk mis-compilation in the face of a language-compliant
user-defined macro that collides with the bare-word counterpart.

We don't have to decorate our attribute usage in .c files,
because there we are in control of making sure no collisions
will interfere.

Signed-off-by: Eric Blake <eblake@redhat.com>