Change the default protocol to newstyle.
nbdkit <= 1.1.28 defaulted to the oldstyle protocol for compatibility
with qemu and libguestfs.  However qemu >= 2.6 can now work with
either protocol and is widely installed.  Also newstyle is required
for newer features such as export names and TLS.  In addition
nbd-client dropped support for oldstyle entirely.

You can select the oldstyle protocol by adding ‘-o’, and it is still
tested.