todo: Reduce thread overhead in parallel mode
Rather than pre-creating the full thread pool for a parallel plugin
(a default of 16 threads, if command line didn't include --thread),
it would be nice to ramp up the number of threads according to
client demand.  A client that always waits for a response does
not need to waste the remaining server threads that will always
be idle.  This in turn would make the test-parallel logs less
verbose.

Signed-off-by: Eric Blake <eblake@redhat.com>