connections: Don't advertise TRIM on readonly connection
A client cannot attempt TRIM on a read-only connection (we already
filter it in connections.c with EROFS); as such, it makes no sense
to query the plugin's .can_trim nor to advertise TRIM support to
the client on such a connection, similar to how we do not
advertise WRITE_ZEROES.

FUA only has defined meanings on write operations, but the NBD
protocol allows us to advertise it on a readonly connection, and to
ignore clients that send it on read (older qemu did so); so for now
we can continue to advertise it regardless of readonly state
(although once we add a separate .can_fua, gating it to writable
connections is also ok).  Meanwhile, although flush is unlikely to
make a difference on a read-only connection, there is no reason
why it shouldn't be advertised regardless of readonly state.

Signed-off-by: Eric Blake <eblake@redhat.com>