maint: Improve ./nbdkit handling of --opt=value
Although useful for in-tree debugging, the ./nbdkit script was not
very nice when it comes to using the '--longopt=value' single-string
form of options, compared to the two-string '--longopt value' form.
Since we used a glob for the multiple spellings of --exportname and
--pidfile, the script would try to treat the next argument literally
instead of recognizing that the argument was already consumed.
Likewise, the script was not recognizing '--filter=foo'.

Since the script already relies on bash, we can exploit bash's
case fallthrough (using ';&' instead of ';;'), to rewrite inline
--filter=foo without having to duplicate logic.

This does NOT try to honor option abbreviations.  If you are testing
that aspect of getopt_long, don't use this script ;)

Signed-off-by: Eric Blake <eblake@redhat.com>
(cherry picked from commit 091ce4024f565a174ccd2aa3206ff431c7883396)