filters: Document lack of ABI guarantee
Guaranteeing a stable filter ABI is extremely difficult: if we
add a new callback but the old filter fails to intercept it,
letting the new callback fall through to the plugin unmodified
may corrupt client-visible state.  Conversely, adding a new
callback generally means we will also want to expand struct
nbdkit_next_ops to let the filter access the new callback, but
a new filter cannot safely access members of the struct not
provided by an older nbdkit.  So beef up the documentation,
and tighten the filter check to gracefully detect incompatible
filters by the only ABI guarantee we really have - the API
version that the filter was compiled against lives in a fixed
location.  We don't need to play games with recording the
struct size, since the API version will change when the struct
size changes.

Add a discussion about ABI/API stability to the plugin docs
while at it, for comparison.

Signed-off-by: Eric Blake <eblake@redhat.com>