filters: cache, cow: Handle bitmap overflow on 32 bit architectures.
When compiling on armv7 it becomes clear from the compiler warnings
that the current code is wrong.

The bitmap has to be allocated in virtual memory, so use size_t to
describe the length of the bitmap.  When changing the length of the
bitmap, compute the new size as an unsigned 64 bit int, and then check
whether or not it is too large to fit into size_t before casting it.