filters: Add log filter
'nbdkit -v' is quite verbose, and traces everything.  Well, actually
it only traces if you also use -f; because if we daemonize, stderr
is redirected to /dev/null.  Sometimes, we want to trace JUST the
client interactions, and it would be nice to trace this even when
run in the background.  In particular, when debugging another
filter, being able to trace what an earlier filter changed can be
quite useful; and having the log filter in place will make it easier
to add testsuite coverage of other filters (this patch includes a
rudimentary test of just the log filter).  Also, it is nice to have
timestamps in the log, in order to see if message interleaving takes
place, and what delays are occurring.

The filter requires a logfile; I debated making the filter fall
back to logging to stderr if no logfile parameter was found, but
while that works for 'nbdkit -f', it fails miserably when we
daemonize (again, part of the argument for having this filter).

Signed-off-by: Eric Blake <eblake@redhat.com>