build: Add unit-testing of internal files
Everything in tests/ is useful for testing our public interface:
nbdkit as a whole, as well as plugins and filters.  But we are
lacking some unit testing of internal functionality, where it
is better to build a standalone binary rather than a full-blown
plugin.  Unit testing belongs best in src/, in part because
cross-directory linking (which would be required if the unit
test lived in tests/) makes automake unhappy.

The first such unit test is of the nbdkit_parse_size() helper
function that is exposed to both plugins and filters.  The test
intentionally has several lines filtered out due to current
weaknesses in the code, which will be fixed later.

Having tests in two different directories means that you now have
to use 'make check -C tests/ TESTS=...' rather than just
'make check TESTS=...' when limiting (re-)testing to just a single
test, but that's not too bad (projects that use gnulib tests are
already familiar with that restriction).

Signed-off-by: Eric Blake <eblake@redhat.com>